More updates for refactor
