--- .env
@@ -1,5 +1,0 @@
-AB2D_DB_HOST=db
-AB2D_DB_PORT=5432
-AB2D_DB_DATABASE=ab2d
-AB2D_DB_USER=ab2d
-AB2D_DB_PASSWORD=ab2d
