+++ api/pom.xml
@@ -18,16 +18,12 @@
     <dependencies>
         <dependency>
             <groupId>gov.cms.ab2d</groupId>
-            <artifactId>domain</artifactId>
+            <artifactId>common</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-security</artifactId>
         </dependency>
         <dependency>
             <groupId>io.springfox</groupId>
