--- api/src/main/java/gov/cms/ab2d/api/service/JobService.java
@@ -1,14 +1,0 @@
-package gov.cms.ab2d.api.service;
-
-import gov.cms.ab2d.domain.Job;
-
-public interface JobService {
-
-    Job createJob(String resourceTypes, String url);
-
-    void cancelJob(String jobId);
-
-    Job getJobByJobID(String jobId);
-
-    Job updateJob(Job job);
-}
