--- api/src/main/java/gov/cms/ab2d/api/service/UserService.java
@@ -1,8 +1,0 @@
-package gov.cms.ab2d.api.service;
-
-import gov.cms.ab2d.domain.User;
-
-public interface UserService {
-
-    User getCurrentUser();
-}
