+++ api/src/main/java/gov/cms/ab2d/api/util/Constants.java
@@ -10,8 +10,6 @@ public final class Constants {
 
     public static final String API_PREFIX = "/api/v1";
 
-    public static final String OPERATION_OUTCOME = "OperationOutcome";
-
     private Constants() {
     }
