+++ common/src/main/java/gov/cms/ab2d/common/util/Constants.java
@@ -1,0 +1,8 @@
+package gov.cms.ab2d.common.util;
+
+public final class Constants {
+
+    private Constants() { }
+
+    public static final String OPERATION_OUTCOME = "OperationOutcome";
+}
