+++ common/src/test/java/gov/cms/ab2d/common/SpringBootApp.java
@@ -1,17 +1,16 @@
-package gov.cms.ab2d.api;
+package gov.cms.ab2d.common;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
-@SpringBootApplication(scanBasePackages = {"gov.cms.ab2d.common", "gov.cms.ab2d.api"})
+@SpringBootApplication(scanBasePackages = "gov.cms.ab2d.common")
 @EntityScan(basePackages = {"gov.cms.ab2d.common.model"})
-@EnableJpaRepositories("gov.cms.ab2d.common.repository")
+@EnableJpaRepositories(basePackages = "gov.cms.ab2d.common.repository")
 public class SpringBootApp {
 
-    public static void main(String[] args) {
+    public static void main(String [] args) {
         SpringApplication.run(SpringBootApp.class, args);
     }
-
 }
