--- domain/src/test/resources/application.properties
@@ -1,8 +1,0 @@
-spring.datasource.driver-class-name=org.h2.Driver
-spring.datasource.url=jdbc:h2:mem:db;DB_CLOSE_ON_EXIT=false
-spring.datasource.username=sa
-spring.datasource.password=sa
-spring.integration.jdbc.initialize-schema=always
-
-spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
-spring.jpa.hibernate.ddl-auto=create-drop
